Return 307 for ApiRootView for ansible-galaxy
If ansible-galaxy gets a 301 or 302 redirect from
a server, when it follows the new Location, it does
not include any authentication info.

Since ansible-galaxy also always appends '/api'
which lacks a trailing '/', there is always at
least one redirect plus the redirect from
$real_url/api -> $real_url.

Which means requests to / to get the ApiRootView
with the available versions and auth url info,
ansible-galaxy doesn't send auth.

But if the redirect is a 307 or 308, it will follow.
So make the $real_url/api -> $real_url redirect a
307. Otherwise ansible-galaxy doesn't work against
Automation Hub.