Make an empty /collections/{ns}/{n} a 404
This is related to https://github.com/ansible/ansible/issues/63210

If pulp returned an empty list for a collection version, galaxy-api
CollectionVersionViewset() would return a 200 and an empty list.
Pulp may return an empty list if the collection doesn't exist.

ansible-galaxy expects a 404 to be returned if the collection
doesn't exist.